+++ mycli/main.py
@@ -77,6 +77,11 @@ Query = namedtuple('Query', ['query', 'successful', 'mutating'])
 
 PACKAGE_ROOT = os.path.abspath(os.path.dirname(__file__))
 
+SUPPORT_INFO = (
+    'Home: http://mycli.net\n'
+    'Bug tracker: https://github.com/dbcli/mycli/issues'
+)
+
 
 class MyCli(object):
 
@@ -561,9 +566,7 @@ class MyCli(object):
         if not self.less_chatty:
             print(' '.join(sqlexecute.server_type()))
             print('mycli', __version__)
-            print('Chat: https://gitter.im/dbcli/mycli')
-            print('Mail: https://groups.google.com/forum/#!forum/mycli-users')
-            print('Home: http://mycli.net')
+            print(SUPPORT_INFO)
             print('Thanks to the contributor -', thanks_picker([author_file, sponsor_file]))
 
         def get_message():
@@ -1083,7 +1086,7 @@ class MyCli(object):
               help='Warn before running a destructive query.')
 @click.option('--local-infile', type=bool,
               help='Enable/disable LOAD DATA LOCAL INFILE.')
-@click.option('--login-path', type=str,
+@click.option('-g', '--login-path', type=str,
               help='Read this path from the login file.')
 @click.option('-e', '--execute',  type=str,
               help='Execute command and quit.')
@@ -1350,6 +1353,9 @@ def read_ssh_config(ssh_config_path):
     try:
         with open(ssh_config_path) as f:
             ssh_config.parse(f)
+    except FileNotFoundError as e:
+        click.secho(str(e), err=True, fg='red')
+        sys.exit(1)
     # Paramiko prior to version 2.7 raises Exception on parse errors.
     # In 2.7 it has become paramiko.ssh_exception.SSHException,
     # but let's catch everything for compatibility
@@ -1358,9 +1364,6 @@ def read_ssh_config(ssh_config_path):
             f'Could not parse SSH configuration file {ssh_config_path}:\n{err} ',
             err=True, fg='red'
         )
-        sys.exit(1)
-    except FileNotFoundError as e:
-        click.secho(str(e), err=True, fg='red')
         sys.exit(1)
     else:
         return ssh_config