+++ changelog.md
@@ -1,9 +1,16 @@
 TODO
 ====
 
+Bug Fixes:
+----------
+* Allow `FileNotFound` exception for SSH config files.
+
+Features:
+---------
+* Add `-g` shortcut to option `--login-path`.
+
 Internal:
 ---------
-
 * Test various host-port combinations in command line arguments